Update README.md
This project is focused on Predicting House Prices using Linear Regression. Predicting house prices is a classic example of a regression problem in machine learning, where the goal is to predict a continuous variable (the price of a house) based on various features (e.g., number of rooms, crime rate, proximity to amenities, etc.).

The Boston Housing Dataset, which is the dataset used in this project, is a well-known dataset in the data science and machine learning community. It contains information about various housing characteristics in Boston, such as the average number of rooms, the crime rate in different neighborhoods, and more. These features are used to predict the price of homes in the area. The dataset is small and contains only 506 samples, which makes it ideal for training beginner-level machine learning models and performing basic analysis.